Add Haversine formula

+++ src/main/java/com/github/dgoldsb/osm/parser/Bounds.java
@@ -1,10 +1,9 @@
 package com.github.dgoldsb.osm.parser;
 
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
-import java.math.BigDecimal;
 
 public record Bounds(
-    @JacksonXmlProperty(localName = "minlat") BigDecimal minimumLatitude,
-    @JacksonXmlProperty(localName = "minlon") BigDecimal minimumLongitude,
-    @JacksonXmlProperty(localName = "maxlat") BigDecimal maximumLatitude,
-    @JacksonXmlProperty(localName = "maxlon") BigDecimal maximumLongitude) {}
+    @JacksonXmlProperty(localName = "minlat") Double minimumLatitude,
+    @JacksonXmlProperty(localName = "minlon") Double minimumLongitude,
+    @JacksonXmlProperty(localName = "maxlat") Double maximumLatitude,
+    @JacksonXmlProperty(localName = "maxlon") Double maximumLongitude) {}
